sql changing

+++ HuliWebShop/src/main/resources/db/migration/V1.0.20230911175237__create_carts_table.sql
@@ -1,4 +1,4 @@
 CREATE TABLE carts (
                        id SERIAL PRIMARY KEY,
-                       user_id BIGINT REFERENCES users(id)
+                       user_id INT REFERENCES users(id)
 );
